Generic ARM fixes (#16994)

+++ src/tests/ngraph_helpers/lpt_ngraph_functions/src/common/fake_quantize_on_weights.cpp
@@ -12,7 +12,7 @@ namespace subgraph {
 FakeQuantizeOnWeights::FakeQuantizeOnWeights() {}
 
 FakeQuantizeOnWeights::FakeQuantizeOnWeights(
-    const size_t quantizationLevel,
+    const uint64_t quantizationLevel,
     const ngraph::Shape& constantShape,
     const std::vector<float>& inputLowValues,
     const std::vector<float>& inputHighValues,
